Connect cart to the user

+++ app/(main)/cart/actions.tsx
@@ -3,20 +3,27 @@
 import { Ingredient } from '@prisma/client';
 import { revalidatePath } from 'next/cache';
 
+import { validateRequest } from '@/app/(auth)/actions';
 import { publicUrls } from '@/config/url';
+import { UnauthorizedError } from '@/lib/errors/UnauthorizedError';
 import { prisma } from '@/prisma/prisma-client';
 
 async function updateOrCreateCart(
   recipeId: number,
   ingredients: Array<Ingredient>,
 ): Promise<{ id: number }> {
+  const { user } = await validateRequest();
+  if (user === null) throw new UnauthorizedError();
+
   const items = ingredients.map(({ id: ingredientId }) => ({ recipeId, ingredientId }));
 
-  const existingCart = await prisma.cart.findFirst();
-  if (existingCart === null) return prisma.cart.create({ data: { items: { create: items } } });
+  const existingCart = await prisma.cart.findFirst({ where: { userId: user.id } });
+  if (existingCart === null) {
+    return prisma.cart.create({ data: { userId: user.id, items: { create: items } } });
+  }
 
   return prisma.cart.update({
-    where: { id: existingCart.id },
+    where: { id: existingCart.id, userId: user.id },
     data: { items: { create: items } },
   });
 }
@@ -32,12 +39,13 @@ export async function addRecipeToCart(
 }
 
 export async function removeRecipeFromCart(recipeId: number): Promise<void> {
-  const existingCart = await prisma.cart.findFirst();
+  const { user } = await validateRequest();
+  if (user === null) throw new UnauthorizedError();
+
+  const existingCart = await prisma.cart.findFirst({ where: { userId: user.id } });
   if (existingCart === null) return;
 
-  await prisma.cartItem.deleteMany({
-    where: { cartId: existingCart.id, recipeId: recipeId },
-  });
+  await prisma.cartItem.deleteMany({ where: { cartId: existingCart.id, recipeId: recipeId } });
 
   revalidatePath(publicUrls.cart);
 }
@@ -46,7 +54,10 @@ export async function removeIngredientFromCart(
   recipeId: number,
   ingredientId: number,
 ): Promise<void> {
-  const existingCart = await prisma.cart.findFirst();
+  const { user } = await validateRequest();
+  if (user === null) throw new UnauthorizedError();
+
+  const existingCart = await prisma.cart.findFirst({ where: { userId: user.id } });
   if (existingCart === null) return;
 
   await prisma.cartItem.deleteMany({
@@ -57,7 +68,10 @@ export async function removeIngredientFromCart(
 }
 
 export async function updateServings(recipeId: number, quantity: number): Promise<void> {
-  const existingCart = await prisma.cart.findFirst();
+  const { user } = await validateRequest();
+  if (user === null) throw new UnauthorizedError();
+
+  const existingCart = await prisma.cart.findFirst({ where: { userId: user.id } });
   if (existingCart === null) return;
 
   await prisma.cartItem.updateMany({
